requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Environment and configuration
python-dotenv==1.0.0

# Web requests
requests==2.31.0

# Data processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2
torch==2.0.1
tensorflow==2.13.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
openai-whisper==20230918

# AI APIs
openai==0.28.1
google-generativeai==0.2.0

# NLP and transformers
transformers==4.33.2
tokenizers==0.13.3

# Feature extraction and analysis
matplotlib==3.7.2
seaborn==0.12.2

# Encryption and security
cryptography==41.0.4

# Utilities
pathlib2==2.3.7
tqdm==4.66.1
argparse==1.4.0

# Vietnamese language processing (optional)
underthesea==6.6.1

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0

# Async support
asyncio==3.4.3

# Additional utilities for ML pipeline
xgboost==1.7.6
catboost==1.2.2

# Audio format support
pydub==0.25.1
ffmpeg-python==0.2.0

# Progress bars and visualization
plotly==5.16.1

# Type checking
typing-extensions==4.7.1

# Concurrent processing
concurrent-futures==3.1.1

# Heroku specific
whitenoise==6.5.0

# Production server
eventlet==0.33.3
